+++ python/tskit/trees.py
@@ -699,6 +699,17 @@ class Tree:
         copy._make_arrays()
         return copy
 
+    # TODO make this method public and document it.
+    # Note that this probably does not cover all corner cases correctly
+    # https://github.com/tskit-dev/tskit/issues/1908
+    def _has_isolated_samples(self):
+        # TODO Is this definition correct for a single-node tree sequence?
+        for root in self.roots:
+            # If the root has no children then it must be a sample
+            if self.left_child(root) == NULL:
+                return True
+        return False
+
     def _make_arrays(self):
         # Store the low-level arrays for efficiency. There's no real overhead
         # in this, because the refer to the same underlying memory as the
@@ -2420,16 +2431,10 @@ class Tree:
             root = self.root
 
         if precision is None:
-            # Quick hack to get sensible default behaviour. We can use the
-            # ``discrete_time`` check when it's implemented.
-            # https://github.com/tskit-dev/tskit/issues/1839
-            time = self.tree_sequence.tables.nodes.time
-            if np.all(time == np.floor(time)):
-                precision = 0
-            else:
-                # 17 decimal digits provides the full precision of an IEEE double,
-                # as defined by DBL_DECIMAL_DIG macro
-                precision = 17
+            # 17 decimal digits provides the full precision of an IEEE double,
+            # as defined by DBL_DECIMAL_DIG macro. If we have discrete time
+            # then write out integer branch lengths.
+            precision = 0 if self.tree_sequence.discrete_time else 17
         include_branch_lengths = (
             True if include_branch_lengths is None else include_branch_lengths
         )
@@ -4511,7 +4516,15 @@ class TreeSequence:
             if tree2.interval.right == right:
                 tree2 = next(trees2, None)
 
-    def _haplotypes_array(self, *, isolated_as_missing, missing_data_character):
+    def _haplotypes_array(
+        self,
+        *,
+        isolated_as_missing=None,
+        missing_data_character=None,
+    ):
+        missing_data_character = (
+            "N" if missing_data_character is None else missing_data_character
+        )
         H = np.empty((self.num_samples, self.num_sites), dtype=np.int8)
         missing_int8 = ord(missing_data_character.encode("ascii"))
         for var in self.variants(isolated_as_missing=isolated_as_missing):
@@ -4548,7 +4561,7 @@ class TreeSequence:
         self,
         *,
         isolated_as_missing=None,
-        missing_data_character="N",
+        missing_data_character=None,
         impute_missing_data=None,
     ):
         """
@@ -4683,10 +4696,10 @@ class TreeSequence:
             as a Python bytes object. Legacy use only.
         :param array_like samples: An array of node IDs for which to generate
             genotypes, or None for all sample nodes. Default: None.
-        :param bool isolated_as_missing: If True, the allele assigned to
+        :param bool isolated_as_missing: If True, the genotype value assigned to
             missing samples (i.e., isolated samples without mutations) is
-            the ``missing_data_character``. If False, missing samples will be
-            assigned the ancestral state.
+            :data:`.MISSING_DATA` (-1). If False, missing samples will be
+            assigned the allele index for the ancestral state.
             Default: True.
         :param tuple alleles: A tuple of strings defining the encoding of
             alleles as integer genotype values. At least one allele must be provided.
@@ -4756,10 +4769,10 @@ class TreeSequence:
             all genotypes are not needed at once, it is usually better to
             access them sequentially using the :meth:`.variants` iterator.
 
-        :param bool isolated_as_missing: If True, the allele assigned to
+        :param bool isolated_as_missing: If True, the genotype value assigned to
             missing samples (i.e., isolated samples without mutations) is
             :data:`.MISSING_DATA` (-1). If False, missing samples will be
-            assigned the ancestral state.
+            assigned the allele index for the ancestral state.
             Default: True.
         :param tuple alleles: A tuple of strings describing the encoding of
             alleles to genotype values. At least one allele must be provided.
@@ -4789,13 +4802,7 @@ class TreeSequence:
             isolated_as_missing=isolated_as_missing, alleles=alleles
         )
 
-    def alignments(
-        self,
-        *,
-        reference_sequence=None,
-        isolated_as_missing=None,
-        missing_data_character=None,
-    ):
+    def alignments(self, *, reference_sequence=None):
         """
         Returns an iterator over the full sequence alignments for the samples in this
         tree sequence. Each alignment ``a`` is a string of length ``L`` where ``L``
@@ -4836,13 +4843,11 @@ class TreeSequence:
         single byte characters, (i.e., variants must be single nucleotide
         polymorphisms, or SNPs).
 
-        If ``isolated_as_missing`` is True (the default), isolated samples without
-        mutations directly above them will be treated as
-        :ref:`missing data<sec_data_model_missing_data>` and will be
-        represented in the string by the ``missing_data_character``. If
-        instead it is set to False, missing data will be assigned the ancestral state
-        (unless they have mutations directly above them, in which case they will take
-        the most recent derived mutational state for that node).
+        .. note:: :ref:`Missing data<sec_data_model_missing_data>` is not
+            currently supported by this method and it will raise a ValueError
+            if called on tree sequences containing isolated samples.
+            See https://github.com/tskit-dev/tskit/issues/1896 for more
+            information.
 
         See also the :meth:`.variants` iterator for site-centric access
         to sample genotypes and :meth:`.haplotypes` for access to sample sequences
@@ -4852,29 +4857,16 @@ class TreeSequence:
             this tree sequence.
         :param str reference_sequence: The reference sequence to fill in
             gaps between sites in the alignments.
-        :param bool isolated_as_missing: If True, the allele assigned to
-            missing samples (i.e., isolated samples without mutations) is
-            the ``missing_data_character``. If False,
-            missing samples will be assigned the ancestral state.
-            Default: True.
-        :param str missing_data_character: A single ascii character that will
-            be used to represent missing data.
-            If any normal allele contains this character, an error is raised.
-            Default: 'N'.
         :rtype: collections.abc.Iterable
         :raises: ValueError
-            if all coordinates in this tree sequence are not discrete, or
-            if the ``missing_data_character`` exists in one of the alleles,
+            if any genome coordinate in this tree sequence is not discrete,
             or if the ``reference_sequence`` is not of the correct length.
-        :raises: TypeError if the ``missing_data_character`` or any of the alleles
-            at a site are not a single ascii character.
+        :raises: TypeError if any of the alleles at a site are not a
+            single ascii character.
         """
         if not self.discrete_genome:
             raise ValueError("sequence alignments only defined for discrete genomes")
 
-        missing_data_character = (
-            "N" if missing_data_character is None else missing_data_character
-        )
         L = int(self.sequence_length)
         a = np.empty(L, dtype=np.int8)
         if reference_sequence is None:
@@ -4897,10 +4889,24 @@ class TreeSequence:
                 )
             ref_bytes = reference_sequence.encode("ascii")
             a[:] = np.frombuffer(ref_bytes, dtype=np.int8)
-        H = self._haplotypes_array(
-            isolated_as_missing=isolated_as_missing,
-            missing_data_character=missing_data_character,
-        )
+
+        # To do this properly we'll have to detect the missing data as
+        # part of a full implementation of alignments in C. The current
+        # definition might not be calling some degenerate cases correctly;
+        # see https://github.com/tskit-dev/tskit/issues/1908
+        #
+        # Note also that this will call the presence of missing data
+        # incorrectly if have a sample isolated over the region (a, b],
+        # and if we have sites at each position from a to b, and at
+        # each site there is a mutation over the isolated sample.
+        if any(tree._has_isolated_samples() for tree in self.trees()):
+            raise ValueError(
+                "Missing data not currently supported in alignments; see "
+                "https://github.com/tskit-dev/tskit/issues/1896 for details."
+                "The current implementation may also incorrectly identify an "
+                "input tree sequence has having missing data."
+            )
+        H = self._haplotypes_array()
         site_pos = self.tables.sites.position.astype(np.int64)
         for h in H:
             a[site_pos] = h
@@ -5293,14 +5299,11 @@ class TreeSequence:
         *,
         wrap_width=60,
         reference_sequence=None,
-        missing_data_character=None,
-        isolated_as_missing=None,
     ):
         """
         Writes the :meth:`.alignments` for this tree sequence to file in
         `FASTA <https://en.wikipedia.org/wiki/FASTA_format>`__ format. The
-        ``reference_sequence``, ``missing_data_character`` and
-        ``isolated_as_missing`` arguments are passed directly to the
+        ``reference_sequence`` argument is passed directly to the
         :meth:`.alignments` method; please its documentation for more details.
 
         .. warning:: The ``reference_sequence`` is currently a mandatory
@@ -5327,8 +5330,14 @@ class TreeSequence:
 
         .. code-block:: python
 
-            ref = "-" * int(ts.sequence_length)
+            ref = "N" * int(ts.sequence_length)
             ts.write_fasta("output.fa", reference_sequence=ref)
+
+        .. note:: :ref:`Missing data<sec_data_model_missing_data>` is not
+            currently supported by this method and it will raise a ValueError
+            if called on tree sequences containing isolated samples.
+            See https://github.com/tskit-dev/tskit/issues/1896 for more
+            information.
 
         :param file_or_path: The file object or path to write the output.
             Paths can be either strings or :class:`python:pathlib.Path` objects.
@@ -5337,22 +5346,18 @@ class TreeSequence:
             to the next line for each sequence, or 0 to turn off line wrapping.
             (Default=60).
         :param str reference_sequence: As for the :meth:`.alignments` method.
-        :param bool isolated_as_missing: As for the :meth:`.alignments` method.
-        :param str missing_data_character: As for the :meth:`.alignments` method.
         """
         text_formats.write_fasta(
             self,
             file_or_path,
             wrap_width=wrap_width,
             reference_sequence=reference_sequence,
-            missing_data_character=missing_data_character,
-            isolated_as_missing=isolated_as_missing,
         )
 
     def as_fasta(self, **kwargs):
         """
-        Return the result of :meth:`.write_fasta` as a string without writing
-        to file. Keyword parameters are as defined in :meth:`.write_fasta`.
+        Return the result of :meth:`.write_fasta` as a string.
+        Keyword parameters are as defined in :meth:`.write_fasta`.
 
         :return: A FASTA encoding of the alignments in this tree sequence as a string.
         :rtype: str
@@ -5369,12 +5374,16 @@ class TreeSequence:
         include_trees=None,
         include_alignments=None,
         reference_sequence=None,
-        missing_data_character=None,
-        isolated_as_missing=None,
     ):
         """
         Returns a `nexus encoding <https://en.wikipedia.org/wiki/Nexus_file>`_
-        of this tree sequence. Tree topologies and branch lengths are listed
+        of this tree sequence. By default, tree topologies are included
+        in the output, and sequence data alignments are included by default
+        if this tree sequence has discrete genome coordinates and one or
+        more sites. Inclusion of these sections can be controlled manually
+        using the ``include_trees`` and ``include_alignments`` parameters.
+
+        Tree topologies and branch lengths are listed
         sequentially in the TREES block and the spatial location of each tree
         encoded within the tree name labels. Specifically, a tree spanning
         the interval :math:`[x, y)`` is given the name ``f"t{x}^{y}"``
@@ -5396,8 +5405,6 @@ class TreeSequence:
         If the ``precision`` argument is provided, genome positions and
         branch lengths are printed out with this many digits of precision.
 
-        .. note:: Missing data is tricky
-
         For example, here is the nexus encoding of a simple tree sequence
         with integer times and genome coordinates with three samples
         and two trees::
@@ -5412,11 +5419,62 @@ class TreeSequence:
               TREE t2^10 = [&R] (n1:2,(n0:1,n2:1):1);
             END;
 
+        If sequence data :meth:`.alignments` are defined for this tree sequence
+        and there is at least one site present, sequence alignment data will also
+        be included by default (this can be suppressed by setting
+        ``include_alignments=False``). For example, this tree sequence has
+        a sequence length of 10 and two variable sites (with a reference
+        genome consisting of "N"s)::
+
+            #NEXUS
+            BEGIN TAXA;
+              DIMENSIONS NTAX=3;
+              TAXLABELS n0 n1 n2;
+            END;
+            BEGIN DATA;
+              DIMENSIONS NCHAR=10;
+              FORMAT DATATYPE=DNA;
+              MATRIX
+                n0 NNGNNNNNNT
+                n1 NNANNNNNNC
+                n2 NNANNNNNNC
+              ;
+            END;
+            BEGIN TREES;
+              TREE t0^10 = [&R] (n0:2,(n1:1,n2:1):1);
+            END;
+
+        .. warning:: The ``reference_sequence`` is currently a mandatory
+            argument if alignment output is required
+            to ensure compatibility with future versions
+            of tskit, which will allow reference sequences to be associated
+            with a tree sequence. In this case, the associated reference
+            will be used by default, if present. See
+            https://github.com/tskit-dev/tskit/issues/1888
+            and
+            https://github.com/tskit-dev/tskit/issues/146 for more details.
+
+        .. note:: :ref:`Missing data<sec_data_model_missing_data>`
+            is not supported for encoding tree topology information
+            as our convention of using trees with multiple roots
+            is not often supported by newick parsers. Thus, the method
+            will raise a ValueError if we try to output trees with
+            multiple roots. Additionally, missing data
+            is not currently supported for alignment data.
+            See https://github.com/tskit-dev/tskit/issues/1896 for more
+            information.
+
         .. seealso: See also the :meth:`.as_nexus` method which will
             return this nexus representation as a string.
 
         :param int precision: The numerical precision with which branch lengths
             and tree positions are printed.
+        :param bool include_trees: True if the tree topology information should
+            be included; False otherwise (default=True).
+        :param bool include_alignments: True if the sequence data alignment information
+            should be included; False otherwise (default=True if sequence alignments
+            are well-defined and the tree sequence contains at least one site).
+        :param str reference_sequence: As for the :meth:`.alignments` method.
         :return: A nexus representation of this :class:`TreeSequence`
         :rtype: str
         """
@@ -5427,15 +5485,15 @@ class TreeSequence:
             include_trees=include_trees,
             include_alignments=include_alignments,
             reference_sequence=reference_sequence,
-            missing_data_character=missing_data_character,
-            isolated_as_missing=isolated_as_missing,
         )
 
     def as_nexus(self, **kwargs):
         """
-        Write the result of :meth:`.as_nexus` directly to file. Keyword
-        parameters are passed to :meth:`.as_nexus`.
+        Return the result of :meth:`.write_nexus` as a string.
+        Keyword parameters are as defined in :meth:`.write_nexus`.
 
+        :return: A nexus encoding of the alignments in this tree sequence as a string.
+        :rtype: str
         """
         buff = io.StringIO()
         self.write_nexus(buff, **kwargs)